Move traceable profiler decorators to their respective components
