+++ src/ai-bundle/src/DependencyInjection/DebugCompilerPass.php
@@ -11,12 +11,12 @@
 
 namespace Symfony\AI\AiBundle\DependencyInjection;
 
-use Symfony\AI\AiBundle\Profiler\TraceableAgent;
-use Symfony\AI\AiBundle\Profiler\TraceableChat;
-use Symfony\AI\AiBundle\Profiler\TraceableMessageStore;
-use Symfony\AI\AiBundle\Profiler\TraceablePlatform;
-use Symfony\AI\AiBundle\Profiler\TraceableStore;
-use Symfony\AI\AiBundle\Profiler\TraceableToolbox;
+use Symfony\AI\Agent\Toolbox\TraceableToolbox;
+use Symfony\AI\Agent\TraceableAgent;
+use Symfony\AI\Chat\TraceableChat;
+use Symfony\AI\Chat\TraceableMessageStore;
+use Symfony\AI\Platform\TraceablePlatform;
+use Symfony\AI\Store\TraceableStore;
 use Symfony\Component\Clock\ClockInterface;
 use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
