+++ src/ai-bundle/src/Profiler/DataCollector.php
@@ -12,8 +12,14 @@
 namespace Symfony\AI\AiBundle\Profiler;
 
 use Symfony\AI\Agent\Toolbox\ToolResult;
+use Symfony\AI\Agent\Toolbox\TraceableToolbox;
+use Symfony\AI\Agent\TraceableAgent;
+use Symfony\AI\Chat\TraceableChat;
+use Symfony\AI\Chat\TraceableMessageStore;
 use Symfony\AI\Platform\Metadata\Metadata;
 use Symfony\AI\Platform\Tool\Tool;
+use Symfony\AI\Platform\TraceablePlatform;
+use Symfony\AI\Store\TraceableStore;
 use Symfony\Bundle\FrameworkBundle\DataCollector\AbstractDataCollector;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
