+++ src/ai-bundle/tests/Profiler/DataCollectorTest.php
@@ -16,15 +16,13 @@ use Symfony\AI\Agent\AgentInterface;
 use Symfony\AI\Agent\MockAgent;
 use Symfony\AI\Agent\Toolbox\Tool\Subagent;
 use Symfony\AI\Agent\Toolbox\ToolboxInterface;
+use Symfony\AI\Agent\Toolbox\TraceableToolbox;
+use Symfony\AI\Agent\TraceableAgent;
 use Symfony\AI\AiBundle\Profiler\DataCollector;
-use Symfony\AI\AiBundle\Profiler\TraceableAgent;
-use Symfony\AI\AiBundle\Profiler\TraceableChat;
-use Symfony\AI\AiBundle\Profiler\TraceableMessageStore;
-use Symfony\AI\AiBundle\Profiler\TraceablePlatform;
-use Symfony\AI\AiBundle\Profiler\TraceableStore;
-use Symfony\AI\AiBundle\Profiler\TraceableToolbox;
 use Symfony\AI\Chat\Chat;
 use Symfony\AI\Chat\InMemory\Store as InMemoryStore;
+use Symfony\AI\Chat\TraceableChat;
+use Symfony\AI\Chat\TraceableMessageStore;
 use Symfony\AI\Platform\Message\Content\Text;
 use Symfony\AI\Platform\Message\Message;
 use Symfony\AI\Platform\Message\MessageBag;
@@ -38,9 +36,11 @@ use Symfony\AI\Platform\Result\StreamResult;
 use Symfony\AI\Platform\Result\TextResult;
 use Symfony\AI\Platform\Tool\ExecutionReference;
 use Symfony\AI\Platform\Tool\Tool;
+use Symfony\AI\Platform\TraceablePlatform;
 use Symfony\AI\Platform\Vector\Vector;
 use Symfony\AI\Store\Document\VectorDocument;
 use Symfony\AI\Store\InMemory\Store;
+use Symfony\AI\Store\TraceableStore;
 use Symfony\Component\Clock\MockClock;
 use Symfony\Component\Clock\MonotonicClock;
 use Symfony\Component\Uid\Uuid;
