+++ src/chat/tests/TraceableMessageStoreTest.php
@@ -9,11 +9,11 @@
  * file that was distributed with this source code.
  */
 
-namespace Symfony\AI\AiBundle\Tests\Profiler;
+namespace Symfony\AI\Chat\Tests;
 
 use PHPUnit\Framework\TestCase;
-use Symfony\AI\AiBundle\Profiler\TraceableMessageStore;
 use Symfony\AI\Chat\InMemory\Store as InMemoryStore;
+use Symfony\AI\Chat\TraceableMessageStore;
 use Symfony\AI\Platform\Message\Message;
 use Symfony\AI\Platform\Message\MessageBag;
 use Symfony\Component\Clock\MonotonicClock;
