+++ src/store/CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 0.7
 ---
 
+ * Add `TraceableStore` profiler decorator moved from AI Bundle
  * Add `RstLoader` and `RstToctreeLoader` for loading RST files and following toctree directives
  * Add pre-query event dispatching for query enhancement before vectorization
  * Add platform-based `Reranker` for cross-encoder reranking via `PlatformInterface`
